Moved fast int32 to char conversion into separate header
